feat: separate agent decorator from server
