+++ python/src/acp_sdk/server/__init__.py
@@ -1,4 +1,5 @@
 from acp_sdk.server.agent import Agent as Agent
+from acp_sdk.server.agent import agent as agent
 from acp_sdk.server.app import create_app as create_app
 from acp_sdk.server.context import Context as Context
 from acp_sdk.server.server import Server as Server
